rollback. don't give pybar socket its own dir.
The `bandit` linter seems to give garbage advice. Following it and
putting PYBAR_SOCKET in its own dir allowed anyone to `rm -rf` it
(because of required UDS permissions). Leaving it in /tmp/ with minimum
permissions is/was actually more secure than what `bandit` recommended,
as only root can modify the socket.